refactor(app): extract profile editing and setup diagnostics support
Move profile-edit flow helpers and setup diagnostics support out of the app facade.

Keep app.rs focused on shared application types, state transitions, and module wiring.
